+++ strutil/example_test.go
@@ -152,8 +152,8 @@ func ExampleFields() {
 }
 
 func ExampleReadField() {
-	fmt.Println(ReadField("Bob    Alice\tJohn Mary", 2, true, " ", "\t"))
-	fmt.Println(ReadField("Bob:::Mary:John:", 3, false, ":"))
+	fmt.Println(ReadField("Bob    Alice\tJohn Mary", 2, true, ' ', '\t'))
+	fmt.Println(ReadField("Bob:::Mary:John:", 3, false, ':'))
 
 	// Output:
 	// John