+++ strutil/strutil_test.go
@@ -156,9 +156,9 @@ func (s *StrUtilSuite) TestReadField(c *C) {
 	c.Assert(ReadField("abc 1234 DEF", 2, true), Equals, "DEF")
 	c.Assert(ReadField("abc 1234 DEF", 3, true), Equals, "")
 
-	c.Assert(ReadField("abc|||||1234||DEF", 1, true, "|"), Equals, "1234")
-	c.Assert(ReadField("abc+1234|DEF", 1, true, "|", "+"), Equals, "1234")
-	c.Assert(ReadField("abc::1234:::DEF:", 5, false, ":"), Equals, "DEF")
+	c.Assert(ReadField("abc|||||1234||DEF", 1, true, '|'), Equals, "1234")
+	c.Assert(ReadField("abc+1234|DEF", 1, true, '|', '+'), Equals, "1234")
+	c.Assert(ReadField("abc::1234:::DEF:", 5, false, ':'), Equals, "DEF")
 }
 
 func (s *StrUtilSuite) TestCopy(c *C) {